optimized the star component and moved the state to RatingPopup
