+++ src/components/RatingPopup/RatingPopup.jsx
@@ -4,6 +4,12 @@ import RatingPopupViewContent from './RatingPopupViewContent';
 
 const RatingPopup = () => {
     const [innerWidth, setInnerWidth] = useState(window.innerWidth);
+    const [randomCodesResponse, setRandomCodesResponse] = useState({});
+    const [ratingStarObj, setRatingStarObj] = useState({
+        firstStar: false,
+        secondStar: false,
+    });
+    const [isLoading, setIsLoading] = useState(true);
 
     const handleResize = useCallback(() => {
         setInnerWidth(window.innerWidth);
@@ -19,12 +25,36 @@ const RatingPopup = () => {
         };
     }, [handleResize]);
 
+    useEffect(() => {
+        async function callRandomCodesAPI() {
+            try {
+                const response = await fetch(import.meta.env.VITE_RANDOM_CODES);
+                const parsedResponse = await response.json();
+                setRandomCodesResponse(parsedResponse);
+            } catch (e) {
+                setRandomCodesResponse({
+                    status: 500,
+                    error: e,
+                });
+            }
+            setIsLoading(false);
+        }
+
+        callRandomCodesAPI();
+    }, []);
+
     return (
         <div className={styles.ratingPopup}>
             <h2 className={styles.ratingPopupH2}>
                 Rate the code file which you see is having the better code style
             </h2>
-            <RatingPopupViewContent innerWidth={innerWidth} />
+            <RatingPopupViewContent
+                innerWidth={innerWidth}
+                isLoading={isLoading}
+                randomCodesResponse={randomCodesResponse}
+                ratingStarObj={ratingStarObj}
+                setRatingStarObj={setRatingStarObj}
+            />
             <div>done button</div>
         </div>
     );
